Refactored Program
